rename model package
