+++ app/src/main/java/com/example/week1/adapter/ContactsAdapter.kt
@@ -1,0 +1,4 @@
+package com.example.week1.adapter
+
+class ContactAdapter {
+}
